LPS-63475 contacts-service - registers upgrade process
